Extract pure functions into traits
